Assistant checkpoint: 创建IssueSaveService和IssueTimeConverter，重构保存逻辑
Assistant generated file changes:
- src/main/java/com/hu/oneclick/server/service/impl/IssueServiceImpl_backup.java: 备份现有的IssueServiceImpl文件
- src/main/java/com/hu/oneclick/server/service/impl/IssueTimeConverter.java: 创建时间转换工具类
- src/main/java/com/hu/oneclick/server/service/impl/IssueSaveService.java: 创建专门处理Issue保存逻辑的服务类
- src/main/java/com/hu/oneclick/server/service/impl/IssueServiceImpl.java: 添加IssueSaveService依赖并简化add和edit方法, 简化add方法，委托给IssueSaveService, 简化edit方法，委托给IssueSaveService, 删除已移到IssueTimeConverter的方法, 修改info方法使用IssueTimeConverter, 修改clone方法使用IssueTimeConverter, 删除已移到IssueTimeConverter的UTC转换方法

---

User prompt:

我们先不要一起做太多， 我们 先创建 IssueSaveService.java， 并且只将与 api/issue/save 的抽取。
 IssueTimeConverter.java - 这个抽离出来。

再创建 IssueCustomFieldProcessor.java - 将自定义字段处理逻辑进一步抽取出来  》 这个先不要做，我们把save 做好以后，现来看。

同时，你备份现在的文件， 因为我们上次修改，一次失败，导致无法回滚，花费太多时间。这次我们要小心，一旦优化失败，我们就可以恢复到目前是正常运的状态。
