+++ src/main/java/com/hu/oneclick/server/service/impl/IssueServiceImpl.java
@@ -68,6 +68,12 @@ public class IssueServiceImpl extends ServiceImpl<IssueDao, Issue> implements Is
 
     private final ViewFilterService viewFilterService;
 
+    @Resource
+    private IssueSaveService issueSaveService;
+
+    @Resource
+    private IssueTimeConverter issueTimeConverter;
+
     // ThreadLocal存储用户时区信息
     private static final ThreadLocal<String> USER_TIMEZONE = new ThreadLocal<>();
 
@@ -103,50 +109,7 @@ public class IssueServiceImpl extends ServiceImpl<IssueDao, Issue> implements Is
     @Override
     @Transactional
     public Issue add(IssueSaveDto dto) {
-        Issue issue = new Issue();
-        BeanUtil.copyProperties(dto, issue);
-
-        // 获取用户时区
-        String userTimezone = TimezoneContext.getUserTimezone();
-
-        // 转换日期到UTC
-        convertDatesToUTC(issue, userTimezone);
-
-        // 处理新增的三个字段：introduced_version, is_legacy, found_after_release
-        Map<String, Object> customFieldMap = new HashMap<>();
-        if (!JSONUtil.isNull(dto.getCustomFieldDatas())) {
-            customFieldMap = JSONUtil.toBean(JSONUtil.toJsonStr(dto.getCustomFieldDatas()), Map.class);
-        }
-
-        // 检查并添加新字段到自定义字段数据中
-        if (dto.getIntroducedVersion() != null) {
-            customFieldMap.put("introduced_version", dto.getIntroducedVersion());
-        }
-        if (dto.getIsLegacy() != null) {
-            customFieldMap.put("is_legacy", dto.getIsLegacy());
-        }
-        if (dto.getFoundAfterRelease() != null) {
-            customFieldMap.put("found_after_release", dto.getFoundAfterRelease());
-        }
-
-        // 保存自定义字段
-        if (!customFieldMap.isEmpty()) {
-            issue.setIssueExpand(JSONUtil.toJsonStr(customFieldMap));
-        }
-
-        // 直接从 DTO 中获取三个版本相关字段
-        if (dto.getIntroducedVersion() != null) {
-            issue.setIntroducedVersion(dto.getIntroducedVersion());
-        }
-        if (dto.getIsLegacy() != null) {
-            issue.setIsLegacy(dto.getIsLegacyAsInt());
-        }
-        if (dto.getFoundAfterRelease() != null) {
-            issue.setFoundAfterRelease(dto.getFoundAfterReleaseAsInt());
-        }
-
-        this.baseMapper.insert(issue);
-        return issue;
+        return issueSaveService.saveNewIssue(dto);
     }
 
     @Override
@@ -157,49 +120,7 @@ public class IssueServiceImpl extends ServiceImpl<IssueDao, Issue> implements Is
             throw new BaseException(StrUtil.format("缺陷查询不到。ID：{} projectId：{}", dto.getId(), dto.getProjectId()));
         }
 
-        Issue issue = new Issue();
-        BeanUtil.copyProperties(dto, issue);
-
-         // 获取用户时区
-        String userTimezone = TimezoneContext.getUserTimezone();
-
-        // 转换日期到UTC
-        convertDatesToUTC(issue, userTimezone);
-
-        // 处理新增的三个字段：introduced_version, is_legacy, found_after_release
-        Map<String, Object> customFieldMap = new HashMap<>();
-        if (!JSONUtil.isNull(dto.getCustomFieldDatas())) {
-            customFieldMap = JSONUtil.toBean(JSONUtil.toJsonStr(dto.getCustomFieldDatas()), Map.class);
-        }
-
-        // 检查并添加新字段到自定义字段数据中
-        if (dto.getIntroducedVersion() != null) {
-            customFieldMap.put("introduced_version", dto.getIntroducedVersion());
-        }
-        if (dto.getIsLegacy() != null) {
-            customFieldMap.put("is_legacy", dto.getIsLegacy());
-        }
-        if (dto.getFoundAfterRelease() != null) {
-            customFieldMap.put("found_after_release", dto.getFoundAfterRelease());
-        }
-
-        // 保存自定义字段
-        if (!customFieldMap.isEmpty()) {
-            issue.setIssueExpand(JSONUtil.toJsonStr(customFieldMap));
-        }
-
-        // 直接从 DTO 中获取三个版本相关字段
-        if (dto.getIntroducedVersion() != null) {
-            issue.setIntroducedVersion(dto.getIntroducedVersion());
-        }
-        if (dto.getIsLegacy() != null) {
-            issue.setIsLegacy(dto.getIsLegacyAsInt());
-        }
-        if (dto.getFoundAfterRelease() != null) {
-            issue.setFoundAfterRelease(dto.getFoundAfterReleaseAsInt());
-        }
-
-        this.baseMapper.updateById(issue);
+        Issue issue = issueSaveService.updateExistingIssue(dto);
 
         // 转换字段格式，确保返回给前端的是字符串格式
         convertFieldsToString(issue);
@@ -207,93 +128,7 @@ public class IssueServiceImpl extends ServiceImpl<IssueDao, Issue> implements Is
         return issue;
     }
 
-    /**
-     * 转换所有日期字段到UTC
-     */
-    private void convertDatesToUTC(Issue issue, String userTimezone) {
-        if (userTimezone == null || userTimezone.isEmpty()) {
-            System.out.println("=== 没有用户时区信息，跳过UTC转换 ===");
-            return;
-        }
-
-        System.out.println("=== 开始UTC转换，用户时区: " + userTimezone + " ===");
-
-        try {
-            TimeZone userTZ = TimeZone.getTimeZone(userTimezone);
-
-            // 转换 createTime（如果是新建时，通常由数据库自动设置）
-            if (issue.getCreateTime() != null) {
-                Date originalTime = issue.getCreateTime();
-                Date utcTime = convertLocalTimeToUTC(originalTime, userTZ);
-                issue.setCreateTime(utcTime);
-                System.out.println("=== createTime转换: " + originalTime + " -> " + utcTime + " ===");
-            }
-
-            // 转换 planFixDate
-            if (issue.getPlanFixDate() != null) {
-                Date originalTime = issue.getPlanFixDate();
-                Date utcTime = convertLocalTimeToUTC(originalTime, userTZ);
-                issue.setPlanFixDate(utcTime);
-                System.out.println("=== planFixDate转换: " + originalTime + " -> " + utcTime + " ===");
-            }
-
-            // 转换 issueExpand 中 attributes 里的日期字段
-            convertAttributesDateFieldsToUTC(issue, userTZ);
-
-        } catch (Exception e) {
-            System.out.println("=== UTC转换失败: " + e.getMessage() + " ===");
-        }
-    }
-
-    private void convertAttributesDateFieldsToUTC(Issue issue, TimeZone userTZ) {
-        if (issue.getIssueExpand() == null || issue.getIssueExpand().isEmpty()) {
-            return;
-        }
-
-        try {
-            JSONObject issueExpandJson = JSONUtil.parseObj(issue.getIssueExpand());
-            JSONArray attributes = issueExpandJson.getJSONArray("attributes");
-
-            if (attributes != null && !attributes.isEmpty()) {
-                for (Object attributeObj : attributes) {
-                    if (attributeObj instanceof JSONObject) {
-                        JSONObject attribute = (JSONObject) attributeObj;
-                        String fieldType = attribute.getStr("fieldType");
-                        String valueData = attribute.getStr("valueData");
-
-                        if ("date".equals(fieldType) && valueData != null && !valueData.isEmpty()) {
-                            try {
-                                Date originalTime = cn.hutool.core.date.DateUtil.parse(valueData);
-                                Date utcTime = convertLocalTimeToUTC(originalTime, userTZ);
-                                attribute.set("valueData", cn.hutool.core.date.DateUtil.format(utcTime, "yyyy-MM-dd HH:mm:ss"));
-                                System.out.println("=== Attribute日期字段转换: " + originalTime + " -> " + utcTime + " ===");
-                            } catch (Exception e) {
-                                System.out.println("=== Attribute日期字段转换失败: " + e.getMessage() + " ===");
-                            }
-                        }
-                    }
-                }
-                issue.setIssueExpand(issueExpandJson.toString());
-            }
-        } catch (Exception e) {
-            System.out.println("=== 处理attributes日期字段失败: " + e.getMessage() + " ===");
-        }
-    }
-
-    /**
-     * 将用户本地时间转换为UTC时间
-     */
-    private Date convertLocalTimeToUTC(Date localTime, TimeZone userTZ) {
-        // 将本地时间解释为用户时区的时间，然后转换为UTC
-        Calendar localCalendar = Calendar.getInstance(userTZ);
-        localCalendar.setTime(localTime);
-
-        // 获取UTC时间
-        Calendar utcCalendar = Calendar.getInstance(TimeZone.getTimeZone("UTC"));
-        utcCalendar.setTimeInMillis(localCalendar.getTimeInMillis());
-
-        return utcCalendar.getTime();
-    }
+    
 
     /**
      * 转换字段格式：确保数据格式正确
@@ -342,7 +177,7 @@ public class IssueServiceImpl extends ServiceImpl<IssueDao, Issue> implements Is
 
         // 获取用户时区并转换UTC时间为用户本地时间
         String userTimezone = TimezoneContext.getUserTimezone();
-        convertUTCToLocalTime(issue, userTimezone);
+        issueTimeConverter.convertUTCToLocalTime(issue, userTimezone);
 
         // 转换字段格式，确保返回给前端的是字符串格式
         convertFieldsToString(issue);
@@ -393,7 +228,7 @@ public class IssueServiceImpl extends ServiceImpl<IssueDao, Issue> implements Is
             issueClone.setIssueStatus("新建");
 
             // 转换时间到UTC
-            convertDatesToUTC(issueClone, userTimezone);
+            issueTimeConverter.convertDatesToUTC(issueClone, userTimezone);
 
             issueList.add(issueClone);
         }
@@ -679,91 +514,5 @@ public class IssueServiceImpl extends ServiceImpl<IssueDao, Issue> implements Is
         System.out.println("=== Duration计算完成 ===");
     }
 
-     /**
-     * 将UTC时间转换为用户本地时间（用于返回给前端显示）
-     */
-    private void convertUTCToLocalTime(Issue issue, String userTimezone) {
-        if (userTimezone == null || userTimezone.isEmpty()) {
-            System.out.println("=== 没有用户时区信息，跳过本地时间转换 ===");
-            return;
-        }
-
-        System.out.println("=== 开始UTC到本地时间转换，用户时区: " + userTimezone + " ===");
-
-        try {
-            TimeZone userTZ = TimeZone.getTimeZone(userTimezone);
-
-            // 转换 createTime
-            if (issue.getCreateTime() != null) {
-                Date originalTime = issue.getCreateTime();
-                Date localTime = convertUTCToLocalTime(originalTime, userTZ);
-                issue.setCreateTime(localTime);
-                System.out.println("=== createTime转换: " + originalTime + " -> " + localTime + " ===");
-            }
-
-            // 转换 planFixDate
-            if (issue.getPlanFixDate() != null) {
-                Date originalTime = issue.getPlanFixDate();
-                Date localTime = convertUTCToLocalTime(originalTime, userTZ);
-                issue.setPlanFixDate(localTime);
-                System.out.println("=== planFixDate转换: " + originalTime + " -> " + localTime + " ===");
-            }
-
-            // 转换 issueExpand 中 attributes 里的日期字段
-            convertAttributesDateFieldsToLocalTime(issue, userTZ);
-
-        } catch (Exception e) {
-            System.out.println("=== UTC到本地时间转换失败: " + e.getMessage() + " ===");
-        }
-    }
-
-    private void convertAttributesDateFieldsToLocalTime(Issue issue, TimeZone userTZ) {
-        if (issue.getIssueExpand() == null || issue.getIssueExpand().isEmpty()) {
-            return;
-        }
-
-        try {
-            JSONObject issueExpandJson = JSONUtil.parseObj(issue.getIssueExpand());
-            JSONArray attributes = issueExpandJson.getJSONArray("attributes");
-
-            if (attributes != null && !attributes.isEmpty()) {
-                for (Object attributeObj : attributes) {
-                    if (attributeObj instanceof JSONObject) {
-                        JSONObject attribute = (JSONObject) attributeObj;
-                        String fieldType = attribute.getStr("fieldType");
-                        String valueData = attribute.getStr("valueData");
-
-                        if ("date".equals(fieldType) && valueData != null && !valueData.isEmpty()) {
-                            try {
-                                Date originalTime = cn.hutool.core.date.DateUtil.parse(valueData);
-                                Date localTime = convertUTCToLocalTime(originalTime, userTZ);
-                                attribute.set("valueData", cn.hutool.core.date.DateUtil.format(localTime, "yyyy-MM-dd HH:mm:ss"));
-                                System.out.println("=== Attribute日期字段转换: " + originalTime + " -> " + localTime + " ===");
-                            } catch (Exception e) {
-                                System.out.println("=== Attribute日期字段转换失败: " + e.getMessage() + " ===");
-                            }
-                        }
-                    }
-                }
-                issue.setIssueExpand(issueExpandJson.toString());
-            }
-        } catch (Exception e) {
-            System.out.println("=== 处理attributes日期字段失败: " + e.getMessage() + " ===");
-        }
-    }
-
-    /**
-     * 将UTC时间转换为用户本地时间
-     */
-    private Date convertUTCToLocalTime(Date utcTime, TimeZone userTZ) {
-        // 将UTC时间解释为UTC时区的时间，然后转换为用户时区
-        Calendar utcCalendar = Calendar.getInstance(TimeZone.getTimeZone("UTC"));
-        utcCalendar.setTime(utcTime);
-
-        // 获取用户本地时间
-        Calendar localCalendar = Calendar.getInstance(userTZ);
-        localCalendar.setTimeInMillis(utcCalendar.getTimeInMillis());
-
-        return localCalendar.getTime();
-    }
+     
 }
